+++ raytracing/bvh.py
@@ -5,10 +5,11 @@ import numpy as np
 import pandas as pd
 
 from raytracing.group import Group
-from raytracing.renderable import Renderable
-from sphere import Sphere
+from raytracing.renderable import Renderable, INF_DISTANCE
+from raytracing.triangle import hits_triangle, Triangle
+from raytracing.sphere import hits_sphere
 import numba as nb
-# import plotly.express as px
+import plotly.express as px
 
 
 np.random.seed(0)
@@ -48,11 +49,12 @@ def get_object_tree_greedy(meshes: list[Renderable], max_objs_per_bb=5, max_dept
         get_object_tree_greedy([meshes[s] for s in split[1]], max_objs_per_bb=max_objs_per_bb, max_depth=max_depth - 1)]
     return Group(groups)
 
-sys.setrecursionlimit(10_000)
+
 def get_object_tree_fast(meshes: list[Renderable], max_objs_per_bb=5, max_depth=1000) -> Group:
-    print(len(meshes))
-    if len(meshes) == 3:
-        pass
+    """
+    TODO: optimizations:    numba implementation (large amount of time is spent on deep nodes with few children
+                            generate bbs once instead of in every function call
+    """
     if len(meshes) <= max_objs_per_bb:
         return Group(meshes)
     bbs = np.stack([mesh.get_bb() for mesh in meshes])
@@ -112,27 +114,247 @@ def vis_group(group: Group):
     fig.show()
 
 
-if __name__ == '__main__':
-    import time
-    s = time.time()
-    res = get_object_tree_fast_nb(30)
-    print(time.time() - s)
-    s = time.time()
-    res = get_object_tree_fast_nb(30)
-    print(time.time() - s)
-    sphere_count = 100
-    sphere_positions = np.random.random((sphere_count, 3))
-    sphere_radius = np.random.random(sphere_count) * 0.05
-    sphere_positions[:, 2] = sphere_radius * 0
-    spheres = [Sphere(pos, r) for pos, r in zip(sphere_positions, sphere_radius)]
+def hits_box(ray_starts, ray_directions, bbs, return_distances=False):
+    bbs_min = bbs[:, 0, :]
+    bbs_max = bbs[:, 1, :]
+    t_min = (bbs_min - ray_starts) / ray_directions
+    t_max = (bbs_max - ray_starts) / ray_directions
+    # negative t means that the ray will not intersect the plane.
+    # Setting to large number will make result in failed boundary check
+    t_min[t_min < 0] = INF_DISTANCE
+    t_max[t_max < 0] = INF_DISTANCE
+    t = t_min
+    mask = t_max < t_min
+    t[mask] = t_max[mask]
+    # t = t[:, :2]
+    hit_points = ray_starts[:, None, :] + ray_directions[:, None, :] * t[:, :, None]
 
-    tree = get_object_tree_greedy(spheres)
+    margin = 0.0001
+    hits = (bbs_min[:, None, :] <= hit_points + margin) & (hit_points <= bbs_max[:, None, :] + margin)
 
-    vis_group(tree)
+    # hits within limits in all dimensions
+    hits_within_limits = hits.all(axis=-1)
+    hits_any_wall = hits_within_limits.any(axis=-1)
+    if return_distances:
+        distances = t * hits_within_limits
+        distances[distances == 0] = INF_DISTANCE
+        # distances[~hits_any_wall] = INF_DISTANCE
+        return hits_any_wall, distances.min(axis=-1)
+    return hits_any_wall
 
-    res = tree.serialize()
-    df = pd.DataFrame({'x': sphere_positions[:, 0], 'y': sphere_positions[:, 1]})
-    fig = px.scatter(df, x='x', y='y')
-    fig.show()
 
-    pass
+def hit_triangle_bvh(rays, bbs, group_child_indexes, children_types, triangles_data):
+    """
+    Vectorized bounding volume hierarchy traversing, handling triangles
+    """
+    ray_starts = rays[0]
+    ray_count = ray_starts.shape[0]
+    ray_directions = rays[1]
+    depth = 128
+
+
+    _ray_starts = ray_starts.copy()
+    _ray_directions = ray_directions.copy()
+    # candidates - objects that are hit and should be explored
+    candidates = np.zeros((ray_count, depth)).astype(int)
+    candidates_min_possible_distance = np.ones_like(candidates) * INF_DISTANCE
+    candidate_lengths = np.ones(ray_count).astype(int)
+    min_hit_distances = np.ones(ray_count) * INF_DISTANCE
+    # min_hit_distances = np.random.random(ray_count) * 40
+
+    ray_indexes = np.arange(ray_count)
+
+    ray_hit_ids = np.ones(ray_count, dtype=int) * -1
+
+    triangles_h = triangles_data[:, 15]
+    triangles_normals = triangles_data[:, 12:15]
+    triangles_Ts = triangles_data[:, :12].reshape(-1, 3, 4)
+    group_history = []
+    for i in range(20000):
+        # id of elements to be checked (both groups and primitives)
+        explored_ids = candidates[np.arange(candidates.shape[0]), candidate_lengths-1]
+        group_history.append(explored_ids)
+        checked_children_ids = group_child_indexes[explored_ids, :]
+        checked_children_types = children_types[explored_ids, :]
+        checked_children_triangle_mask = (checked_children_types == Triangle.get_type_id())[:, 0]
+        # ray_hit_ids[ray_indexes[checked_children_triangle_mask]] = i
+        checked_children_triangle_indexes = np.where(checked_children_triangle_mask)[0]
+        checked_children_triangle_ids = checked_children_ids[checked_children_triangle_mask, :]
+        checked_children_triangle_ids_0 = checked_children_triangle_ids[:, 0]
+        checked_children_triangle_ids_1 = checked_children_triangle_ids[:, 1]
+        if (checked_children_triangle_ids_0).any():
+            # ray_hit_ids[ray_indexes[checked_children_triangle_mask]] = i
+            pass
+        triangle_0_distances = hits_triangle(ray_starts[checked_children_triangle_mask],
+                                         ray_directions[checked_children_triangle_mask],
+                                         triangles_Ts[checked_children_triangle_ids_0, :, :],
+                                         triangles_normals[checked_children_triangle_ids_0, :],
+                                         triangles_h[checked_children_triangle_ids_0])
+        multi_element_group_mask = checked_children_triangle_ids_1 != -1
+        multi_element_group_indexes = np.where(multi_element_group_mask)[0]
+        triangle_1_distances = hits_triangle(ray_starts[checked_children_triangle_indexes[multi_element_group_mask]],
+                                         ray_directions[checked_children_triangle_indexes[multi_element_group_mask]],
+                                         triangles_Ts[checked_children_triangle_ids_1[multi_element_group_mask], :, :],
+                                         triangles_normals[checked_children_triangle_ids_1[multi_element_group_mask], :],
+                                         triangles_h[checked_children_triangle_ids_1[multi_element_group_mask]])
+        triangle_0_closer = triangle_0_distances < min_hit_distances[checked_children_triangle_indexes]
+        min_hit_distances[checked_children_triangle_indexes[triangle_0_closer]] = triangle_0_distances[triangle_0_closer]
+        ray_hit_ids[ray_indexes[checked_children_triangle_indexes[triangle_0_closer]]] = checked_children_triangle_ids_0[triangle_0_closer]
+
+        triangle_1_closer = triangle_1_distances < min_hit_distances[checked_children_triangle_indexes[multi_element_group_mask]]
+        min_hit_distances[checked_children_triangle_indexes[multi_element_group_indexes[triangle_1_closer]]] = triangle_1_distances[triangle_1_closer]
+        ray_hit_ids[ray_indexes[checked_children_triangle_indexes[multi_element_group_indexes[triangle_1_closer]]]] = (
+            checked_children_triangle_ids_1)[multi_element_group_indexes[triangle_1_closer]]
+        if (triangle_0_distances < INF_DISTANCE).any() or (triangle_1_distances < INF_DISTANCE).any():
+            pass
+
+        checked_children_groups_mask = (checked_children_types == Group.get_type_id())[:, 0]
+        checked_children_groups_indexes = np.where(checked_children_groups_mask)[0]
+        checked_children_group_ids = checked_children_ids[checked_children_groups_mask, :]
+        checked_children_group_ids_0 = checked_children_group_ids[:, 0]
+        checked_children_group_ids_1 = checked_children_group_ids[:, 1]
+
+        group_0_hits, group_0_distances = hits_box(ray_starts[checked_children_groups_mask, :],
+                                                   ray_directions[checked_children_groups_mask, :],
+                                                   bbs[checked_children_group_ids_0, :, :],
+                                                   return_distances=True)
+        # group_0_hits[:] = 1
+        # group_0_distances[:] = 0
+        group_1_hits, group_1_distances = hits_box(ray_starts[checked_children_groups_mask, :],
+                                                   ray_directions[checked_children_groups_mask, :],
+                                                   bbs[checked_children_group_ids_1, :, :],
+                                                   return_distances=True)
+        # group_1_hits[:] = 1
+        # group_1_distances[:] = 0
+        group_0_hits[group_0_distances > min_hit_distances[checked_children_groups_mask]] = 0
+        group_1_hits[group_1_distances > min_hit_distances[checked_children_groups_mask]] = 0
+        # BOTH GROUPS HIT
+        both_groups_hit_mask = group_0_hits & group_1_hits
+        both_groups_hit_indexes = np.where(both_groups_hit_mask)[0]
+        group_1_closer = group_1_distances[both_groups_hit_mask] < group_0_distances[both_groups_hit_mask]
+        swap_indexes = both_groups_hit_indexes[group_1_closer]
+        swap_id_0 = checked_children_group_ids_0[swap_indexes]
+        checked_children_group_ids_0[swap_indexes] = checked_children_group_ids_1[swap_indexes]
+        checked_children_group_ids_1[swap_indexes] = swap_id_0
+
+        candidate_lengths -= 1
+
+        candidates[checked_children_groups_indexes[group_1_hits], candidate_lengths[checked_children_groups_indexes[group_1_hits]]] = checked_children_group_ids_1[group_1_hits]
+        candidate_lengths[checked_children_groups_indexes[group_1_hits]] += 1
+
+        candidates[checked_children_groups_indexes[group_0_hits], candidate_lengths[checked_children_groups_indexes[group_0_hits]]] = checked_children_group_ids_0[group_0_hits]
+        candidate_lengths[checked_children_groups_indexes[group_0_hits]] += 1
+
+        candidate_mask = candidate_lengths > 0
+        if not candidate_mask.any():
+            break
+        ray_starts = ray_starts[candidate_mask, :]
+        ray_directions = ray_directions[candidate_mask, :]
+        ray_indexes = ray_indexes[candidate_mask]
+        candidates = candidates[candidate_mask, :]
+        candidate_lengths = candidate_lengths[candidate_mask]
+        min_hit_distances = min_hit_distances[candidate_mask]
+    print(i)
+    hits = np.zeros(ray_count)
+    hits[ray_indexes] = 1
+    return ray_hit_ids
+
+
+def hit_sphere_bvh(rays, bbs, group_child_indexes, children_types, spheres_data):
+    """
+    Vectorized bounding volume hierarchy traversing
+    """
+    ray_starts = rays[0]
+    ray_count = ray_starts.shape[0]
+    ray_directions = rays[1]
+    depth = 128
+
+    # candidates - objects that are hit and should be explored
+    candidates = np.zeros((ray_count, depth)).astype(int)
+    candidate_lengths = np.ones(ray_count).astype(int)
+    min_hit_distances = np.ones(ray_count) * INF_DISTANCE
+    # min_hit_distances = np.random.random(ray_count) * 40
+
+    ray_indexes = np.arange(ray_count)
+
+    ray_hit_ids = np.ones(ray_count, dtype=int) * -1
+    spheres_pos = spheres_data[:, :3]
+    spheres_r = spheres_data[:, 3]
+    for i in range(20000):
+        # id of elements to be checked (both groups and primitives)
+        explored_ids = candidates[np.arange(candidates.shape[0]), candidate_lengths-1]
+
+        checked_children_ids = group_child_indexes[explored_ids, :]
+        checked_children_types = children_types[explored_ids, :]
+
+        checked_children_sphere_mask = (checked_children_types == ENTITY_TYPE_MAP['sphere'])[:, 0]
+        checked_children_sphere_indexes = np.where(checked_children_sphere_mask)[0]
+        checked_children_sphere_ids = checked_children_ids[checked_children_sphere_mask, :]
+        checked_children_sphere_ids_0 = checked_children_sphere_ids[:, 0]
+        checked_children_sphere_ids_1 = checked_children_sphere_ids[:, 1]
+
+        sphere_0_distances = hits_sphere(ray_starts[checked_children_sphere_mask],
+                                         ray_directions[checked_children_sphere_mask],
+                                         spheres_pos[checked_children_sphere_ids_0, :],
+                                         spheres_r[checked_children_sphere_ids_0])
+        multi_element_group_mask = checked_children_sphere_ids_1 != -1
+        multi_element_group_indexes = np.where(multi_element_group_mask)[0]
+        sphere_1_distances = hits_sphere(ray_starts[checked_children_sphere_indexes[multi_element_group_mask]],
+                                         ray_directions[checked_children_sphere_indexes[multi_element_group_mask]],
+                                         spheres_pos[checked_children_sphere_ids_1[multi_element_group_mask], :],
+                                         spheres_r[checked_children_sphere_ids_1[multi_element_group_mask]])
+        sphere_0_closer = sphere_0_distances < min_hit_distances[checked_children_sphere_indexes]
+        min_hit_distances[checked_children_sphere_indexes[sphere_0_closer]] = sphere_0_distances[sphere_0_closer]
+        ray_hit_ids[ray_indexes[checked_children_sphere_indexes[sphere_0_closer]]] = checked_children_sphere_ids_0[sphere_0_closer]
+
+        sphere_1_closer = sphere_1_distances < min_hit_distances[checked_children_sphere_indexes[multi_element_group_mask]]
+        min_hit_distances[checked_children_sphere_indexes[multi_element_group_indexes[sphere_1_closer]]] = sphere_1_distances[sphere_1_closer]
+        ray_hit_ids[ray_indexes[checked_children_sphere_indexes[multi_element_group_indexes[sphere_1_closer]]]] = (
+            checked_children_sphere_ids_1)[multi_element_group_indexes[sphere_1_closer]]
+
+        checked_children_groups_mask = (checked_children_types == ENTITY_TYPE_MAP['group'])[:, 0]
+        checked_children_groups_indexes = np.where(checked_children_groups_mask)[0]
+        checked_children_group_ids = checked_children_ids[checked_children_groups_mask, :]
+        checked_children_group_ids_0 = checked_children_group_ids[:, 0]
+        checked_children_group_ids_1 = checked_children_group_ids[:, 1]
+
+        group_0_hits, group_0_distances = hits_box(ray_starts[checked_children_groups_mask, :],
+                                                   ray_directions[checked_children_groups_mask, :],
+                                                   bbs[checked_children_group_ids_0, :, :],
+                                                   return_distances=True)
+        group_1_hits, group_1_distances = hits_box(ray_starts[checked_children_groups_mask, :],
+                                                   ray_directions[checked_children_groups_mask, :],
+                                                   bbs[checked_children_group_ids_1, :, :],
+                                                   return_distances=True)
+        group_0_hits[group_0_distances > min_hit_distances[checked_children_groups_mask]] = 0
+        group_1_hits[group_1_distances > min_hit_distances[checked_children_groups_mask]] = 0
+        # BOTH GROUPS HIT
+        both_groups_hit_mask = group_0_hits & group_1_hits
+        both_groups_hit_indexes = np.where(both_groups_hit_mask)[0]
+        group_1_closer = group_1_distances[both_groups_hit_mask] < group_0_distances[both_groups_hit_mask]
+        swap_indexes = both_groups_hit_indexes[group_1_closer]
+        swap_id_0 = checked_children_group_ids_0[swap_indexes]
+        checked_children_group_ids_0[swap_indexes] = checked_children_group_ids_1[swap_indexes]
+        checked_children_group_ids_1[swap_indexes] = swap_id_0
+
+        candidate_lengths -= 1
+
+        candidates[checked_children_groups_indexes[group_1_hits], candidate_lengths[checked_children_groups_indexes[group_1_hits]]] = checked_children_group_ids_1[group_1_hits]
+        candidate_lengths[checked_children_groups_indexes[group_1_hits]] += 1
+
+        candidates[checked_children_groups_indexes[group_0_hits], candidate_lengths[checked_children_groups_indexes[group_0_hits]]] = checked_children_group_ids_0[group_0_hits]
+        candidate_lengths[checked_children_groups_indexes[group_0_hits]] += 1
+
+        candidate_mask = candidate_lengths > 0
+        if not candidate_mask.any():
+            break
+        ray_starts = ray_starts[candidate_mask, :]
+        ray_directions = ray_directions[candidate_mask, :]
+        ray_indexes = ray_indexes[candidate_mask]
+        candidates = candidates[candidate_mask, :]
+        candidate_lengths = candidate_lengths[candidate_mask]
+        min_hit_distances = min_hit_distances[candidate_mask]
+    print(i)
+    hits = np.zeros(ray_count)
+    hits[ray_indexes] = 1
+    return ray_hit_ids
