+++ raytracing/sphere.py
@@ -1,7 +1,6 @@
 import numpy as np
 
-from raytracing.renderable import Renderable
-from models.primitives.sphere import Sphere as SphereMesh
+from raytracing.renderable import Renderable, INF_DISTANCE
 
 
 class Sphere(Renderable):
@@ -16,3 +15,21 @@ class Sphere(Renderable):
 
     def serialize(self):
         return np.concatenate((self.pos, [self.radius]))
+
+
+def hits_sphere(ray_starts, ray_directions, spheres_pos, spheres_r):
+    if len(ray_starts) == 0:
+        return spheres_r
+    dp = spheres_pos - ray_starts
+    a = (ray_directions * ray_directions).sum(axis=1)
+    b = - 2 * (dp * ray_directions).sum(axis=1)
+    c = (dp * dp).sum(axis=1) - spheres_r * spheres_r
+    delta = b * b - 4 * a * c
+    distances = np.ones_like(spheres_r, dtype=float) * INF_DISTANCE
+    # normals = np.zeros_like(ray_starts, dtype=float)
+    hits = delta > 0
+    hit_distances = (-b[hits] - np.sqrt(delta[hits])) / (2 * a[hits])
+    distances[hits] = hit_distances
+    # normals[hits] = ray_starts[hits, :] + ray_directions[hits, :] * hit_distances[:, None]
+    return distances
+
